Fixed c2c255f21a10e51d6b1d1bc0b542a0a3091ed09b
It only worked with integers. Now works with floating point numbers.

+++ easy/02102012.rb
@@ -1,9 +1,9 @@
 #!/usr/bin/ruby
 
 puts "Mass? "
-mass = gets.chomp.to_i
+mass = gets.chomp.to_f
 
 puts "Acceleration? "
-acceleration = gets.chomp.to_i
+acceleration = gets.chomp.to_f
 
 puts "Force = " + (mass * acceleration).to_s
